refactor(oxfmt): Fix up directory structure (#16487)
Until now, most files were stored flat.

From now on:

- `cli/`: CLI-specific code (command, format, init, reporter, result, service,
walk)
- `core/`: Shared formatting logic (format, support, external_formatter)
- `lsp/`: LSP related files

```
src
├── cli
│   ├── command.rs
│   ├── format.rs
│   ├── init.rs
│   ├── mod.rs
│   ├── reporter.rs
│   ├── result.rs
│   ├── service.rs
│   └── walk.rs
├── core
│   ├── external_formatter.rs
│   ├── format.rs
│   ├── mod.rs
│   └── support.rs
├── lib.rs
├── lsp
│   └── mod.rs
├── main.rs
└── main_napi.rs
```
