+++ apps/oxfmt/src/cli/command.rs
