+++ apps/oxfmt/src/cli/mod.rs
@@ -1,0 +1,12 @@
+pub(crate) mod command;
+mod format;
+mod init;
+mod reporter;
+mod result;
+mod service;
+mod walk;
+
+pub use command::format_command;
+pub use format::FormatRunner;
+pub use init::{init_miette, init_tracing};
+pub use result::CliRunResult;
