+++ apps/oxfmt/src/cli/reporter.rs
