+++ apps/oxfmt/src/cli/result.rs
