+++ apps/oxfmt/src/core/mod.rs
@@ -1,0 +1,11 @@
+mod format;
+mod support;
+
+#[cfg(feature = "napi")]
+mod external_formatter;
+
+pub use format::{FormatResult, SourceFormatter};
+pub use support::FormatFileSource;
+
+#[cfg(feature = "napi")]
+pub use external_formatter::{ExternalFormatter, JsFormatEmbeddedCb, JsFormatFileCb};
