--- apps/oxfmt/src/prettier_plugins/mod.rs
@@ -1,3 +1,0 @@
-mod external_formatter;
-
-pub use external_formatter::{ExternalFormatter, JsFormatEmbeddedCb, JsFormatFileCb};
